+++ src/main/scala/mesosphere/marathon/FeaturesConf.scala
@@ -5,12 +5,26 @@ import org.rogach.scallop.{ScallopConf, ValueConverter}
 trait FeaturesConf extends ScallopConf {
   private val setStringParser: ValueConverter[Set[String]] = implicitly[ValueConverter[String]].map(parseList(_))
 
-  private val deprecatedFeatureParser: ValueConverter[DeprecatedFeatureSet] = setStringParser.map { values =>
-    val parsed = values.map { key =>
-      (key, DeprecatedFeatures.all.find(_.key == key))
-    }
+  private val deprecatedFeatureParser: ValueConverter[DeprecatedFeatureConfig] = setStringParser.map { values =>
+    val DisabledRegex = "^disable_(.+)$".r
+    val parsed = values
+      .iterator
+      .map { unparsedKey =>
+        val (key, enabled) = unparsedKey match {
+          case DisabledRegex(key) => key -> false
+          case key => key -> true
+        }
 
-    val unknown = parsed.collect { case (key, None) => key }
+        DeprecatedFeatures.all.find(_.key == key) match {
+          case Some(df) =>
+            Right(df -> enabled)
+          case None =>
+            Left(key)
+        }
+      }
+      .toSeq
+
+    val unknown = parsed.collect { case Left(key) => key }
     require(
       unknown.isEmpty,
       s"Unknown deprecated features specified: ${unknown.mkString(", ")}. Available deprecated features are:\n\n${DeprecatedFeatures.description}" +
@@ -18,8 +32,8 @@ trait FeaturesConf extends ScallopConf {
         "If you recently upgraded, you should downgrade to the old Marathon version and remove the deprecated " +
         "feature(s) in question, ensuring that your cluster continues to function without it."
     )
-    val dfs = parsed.collect { case (_, Some(df)) => df }
-    DeprecatedFeatureSet(BuildInfo.version, dfs)
+    val dfSettings = parsed.collect { case Right(dfSetting) => dfSetting }.toMap
+    DeprecatedFeatureConfig(BuildInfo.version, dfSettings)
   }
 
   lazy val features = opt[Set[String]](
@@ -31,16 +45,16 @@ trait FeaturesConf extends ScallopConf {
     validate = validateFeatures
   )(setStringParser)
 
-  lazy val deprecatedFeatures = opt[DeprecatedFeatureSet](
+  lazy val deprecatedFeatures = opt[DeprecatedFeatureConfig](
     "deprecated_features",
     descr = "A comma-separated list of deprecated features to continue to enable",
     required = false,
-    default = Some(DeprecatedFeatureSet(BuildInfo.version, Set.empty)),
+    default = Some(DeprecatedFeatureConfig(BuildInfo.version, Map.empty)),
     noshort = true,
     validate = { dfs => dfs.isValid() })(deprecatedFeatureParser)
 
   def availableFeatures: Set[String] = features()
-  def availableDeprecatedFeatures: DeprecatedFeatureSet = deprecatedFeatures()
+  def availableDeprecatedFeatures: DeprecatedFeatureConfig = deprecatedFeatures()
 
   def isFeatureSet(name: String): Boolean = availableFeatures.contains(name)