+++ src/main/scala/mesosphere/marathon/api/v2/AppNormalization.scala
@@ -2,7 +2,7 @@ package mesosphere.marathon
 package api.v2
 
 import mesosphere.marathon.raml._
-import mesosphere.marathon.state.{AbsolutePathId, FetchUri, PathId}
+import mesosphere.marathon.state.{AbsolutePathId, FetchUri, PathId, ResourceRole}
 import mesosphere.marathon.stream.Implicits._
 import mesosphere.mesos.ResourceMatcher.Role
 
@@ -202,10 +202,17 @@ object AppNormalization {
       else c
     }
 
-  def sanitizeAcceptedResourceRoles(app: App, effectiveRole: String): Option[Set[String]] =
+  def sanitizeAcceptedResourceRoles(app: App, effectiveRole: String): Option[Set[String]] = {
     app.acceptedResourceRoles.map { roles =>
-      roles.filter(role => role == "*" || role == effectiveRole)
+      val sanitized = roles.filter(role => role == "*" || role == effectiveRole)
+
+      // This method is only called when [[DeprecatedFeatures.sanitizeAcceptedResourceRoles]] is ON. In this
+      // case we not only filter out invalid roles, but also fallback to the default (*) one. Note that acceptedResourceRoles
+      // is about reservations and NOT allocation, so the default one is (*) and not (--mesos_role)
+      if (sanitized.isEmpty) Set(ResourceRole.Unreserved)
+      else sanitized
     }
+  }
 
   def maybeDropPortMappings(c: Container, networks: Seq[Network]): Container =
     // empty networks Seq defaults to host-mode later on, so consider it now as indicating host-mode networking
@@ -297,7 +304,26 @@ object AppNormalization {
     )
   }
 
-  def apply(config: Config): Normalization[App] = Normalization { app =>
+/** 
+ * This is a partial normalization, which currently only normalizes the [[App.role]] and [[App.acceptedResourceRoles]] fields.
+ * We do this because app validation is relying on these fields being set correctly.
+ */
+  def forPreValidation(config: Config): Normalization[App] = Normalization{ app =>
+    val role = app.role.getOrElse(config.defaultRole)
+
+    // sanitize accepted resource roles if enabled
+    val acceptedResourceRoles =
+      if (config.sanitizeAcceptedResourceRoles) {
+        sanitizeAcceptedResourceRoles(app, role)
+      } else app.acceptedResourceRoles
+
+    app.copy(
+      role = Some(role),
+      acceptedResourceRoles = acceptedResourceRoles
+    )
+  }
+
+  def forPostValidation(config: Config): Normalization[App] = Normalization { app =>
     val networks = Networks(config, Some(app.networks)).normalize.networks.filter(_.nonEmpty).getOrElse(DefaultNetworks)
     NetworkNormalization.requireContainerNetworkNameResolution(networks)
     val container = NetworkedContainer(Some(networks), app.container).normalize.container
@@ -310,22 +336,17 @@ object AppNormalization {
     // requirePorts only applies for host-mode networking
     val requirePorts = networks.find(_.mode != NetworkMode.Host).fold(app.requirePorts)(_ => false)
 
-    val role = app.role.getOrElse(config.defaultRole)
-
-    // sanitize accepted resource roles if enabled
-    val acceptedResourceRoles =
-      if (config.sanitizeAcceptedResourceRoles) {
-        sanitizeAcceptedResourceRoles(app, role)
-      } else app.acceptedResourceRoles
-
     app.copy(
       container = container,
       networks = networks,
       unreachableStrategy = app.unreachableStrategy.orElse(Option(defaultUnreachable)),
       requirePorts = requirePorts,
-      role = Some(role),
-      acceptedResourceRoles = acceptedResourceRoles
     )
+  }
+
+  def apply(config: Config): Normalization[App] = Normalization { app =>
+    val preNormalized = forPreValidation(config).normalized(app)
+    forPostValidation(config).normalized(preNormalized)
   }
 
   /** dynamic app normalization configuration, useful for migration and/or testing */