+++ src/main/scala/mesosphere/marathon/util/RoleSettings.scala
@@ -13,24 +13,27 @@ import mesosphere.marathon.state.{RootGroup, AbsolutePathId}
   * @param validRoles List of valid roles
   * @param defaultRole The default role to use if no role is specified on the service. Defaults to '*'
   */
-case class RoleSettings(validRoles: Set[Role], defaultRole: Role) {
+case class RoleSettings(validRoles: Set[Role], defaultRole: Role, previousRole: Option[String] = None, forceRoleUpdate: Boolean = false) {
   require(validRoles(defaultRole))
 }
 
 object RoleSettings extends StrictLogging {
+
   /**
     * Returns the role settings for the service with the specified ID, based on the top-level group and the global config
     *
     * @param config Global config to provide defaultMesos role
     * @param servicePathId The absolute pathId of the affected runSpec, used to determine a possible top-level group role
     * @param rootGroup The root group, used to access possible top-level groups and their settings
+    * @param forceRoleUpdate For resident services, an update to the role may leave instances with their old roles and therefore requires an explicit force update
     *
     * @return A data set with valid roles, default role and a flag if the role should be enforced
     */
   def forService(
     config: MarathonConf,
     servicePathId: AbsolutePathId,
-    rootGroup: RootGroup): RoleSettings = {
+    rootGroup: RootGroup,
+    forceRoleUpdate: Boolean): RoleSettings = {
     val defaultRole = config.mesosRole()
 
     if (servicePathId.parent.isRoot) {
@@ -40,6 +43,8 @@ object RoleSettings extends StrictLogging {
       val topLevelGroupPath = servicePathId.rootPath
       val topLevelGroupRole = topLevelGroupPath.root
       val topLevelGroup = rootGroup.group(topLevelGroupPath)
+
+      val oldServiceRole = rootGroup.runSpec(servicePathId).map(_.role)
 
       val defaultForEnforceFromConfig: Boolean = false // TODO: Use value from config instead
       val enforceRole = topLevelGroup.fold(defaultForEnforceFromConfig)(_.enforceRole)
@@ -57,14 +62,18 @@ object RoleSettings extends StrictLogging {
 
       if (enforceRole) {
         // With enforceRole, we only allow the service to use the group-role or an existing role
-        RoleSettings(validRoles = Set(topLevelGroupRole) ++ maybeExistingRole, defaultRole = topLevelGroupRole)
+        RoleSettings(validRoles = Set(topLevelGroupRole) ++ maybeExistingRole, defaultRole = topLevelGroupRole, previousRole = oldServiceRole, forceRoleUpdate = forceRoleUpdate)
       } else {
         // Without enforce role, we allow default role, group role and already existing role
         // The default role depends on the config parameter
         val defaultRoleToUse = if (defaultForEnforceFromConfig) topLevelGroupRole else defaultRole
-        RoleSettings(validRoles = Set(defaultRole, topLevelGroupRole) ++ maybeExistingRole, defaultRole = defaultRoleToUse)
+        RoleSettings(validRoles = Set(defaultRole, topLevelGroupRole) ++ maybeExistingRole, defaultRole = defaultRoleToUse, previousRole = oldServiceRole, forceRoleUpdate = forceRoleUpdate)
       }
     }
   }
 
+  def residentRoleChangeWarningMessage(previousRole: String, newRole: String): String =
+    "It is not possible to change the role for existing reservations. If you proceed with this change, all existing " +
+      "instances will continue to run under the previous role, " + previousRole + ". Only new instances will be " +
+      "allocated with the new role, " + newRole + ". In order to continue, retry your request with ?force=true"
 }