+++ src/test/scala/mesosphere/marathon/api/v2/PodsResourceTest.scala
@@ -25,7 +25,7 @@ import mesosphere.marathon.raml.{EnvVarSecret, ExecutorResources, FixedPodScalin
 import mesosphere.marathon.state.PathId._
 import mesosphere.marathon.state._
 import mesosphere.marathon.test.{GroupCreation, JerseyTest, Mockito, SettableClock}
-import mesosphere.marathon.util.SemanticVersion
+import mesosphere.marathon.util.{RoleSettings, SemanticVersion}
 import play.api.libs.json._
 
 import scala.collection.immutable.Seq
@@ -462,6 +462,79 @@ class PodsResourceTest extends AkkaUnitTest with Mockito with JerseyTest {
       withClue(s"response body: ${response.getEntity}") {
         response.getStatus should be(422)
         response.getEntity.toString should include("Resident pods cannot have the role *")
+      }
+    }
+
+    "fail to update a pod with a persistent volume without force parameter" in {
+      implicit val podSystem = mock[PodManager]
+      val f = Fixture(configArgs = Seq("--mesos_role", "foo"))
+
+      val pathId = "/foo/mypod".toAbsolutePath
+      val existingPod = PodDefinition(id = pathId, role = "*")
+
+      f.prepareGroup("/foo", Map(pathId.asAbsolutePath -> existingPod))
+
+      podSystem.update(any, eq(false)).returns(Future.successful(DeploymentPlan.empty))
+
+      val podSpecJsonWithPersistentVolume =
+        """
+          | { "id": "/foo/mypod",
+          |   "role": "foo",
+          |   "containers": [ {
+          |     "name": "dataapp",
+          |     "resources": { "cpus": 0.03, "mem": 64 },
+          |     "image": { "kind": "DOCKER", "id": "busybox" },
+          |     "exec": { "command": { "shell": "sleep 1" } },
+          |     "volumeMounts": [ { "name": "pst", "mountPath": "pst1", "readOnly": false } ]
+          |   } ],
+          |   "volumes": [ {
+          |     "name": "pst",
+          |     "persistent": { "type": "root", "size": 10 }
+          |   } ] }
+        """.stripMargin
+
+      val response = asyncRequest { r =>
+        f.podsResource.update("/foo/mypod", podSpecJsonWithPersistentVolume.getBytes(), force = false, f.auth.request, r)
+      }
+      withClue(s"response body: ${response.getEntity}") {
+        response.getStatus should be(422)
+        response.getEntity.toString should include(RoleSettings.residentRoleChangeWarningMessage("*", "foo"))
+      }
+    }
+
+    "update a pod with a persistent volume with force parameter" in {
+      implicit val podSystem = mock[PodManager]
+      val f = Fixture(configArgs = Seq("--mesos_role", "foo"))
+
+      val pathId = "/foo/mypod".toAbsolutePath
+      val existingPod = PodDefinition(id = pathId, role = "*")
+
+      f.prepareGroup("/foo", Map(pathId.asAbsolutePath -> existingPod))
+
+      podSystem.update(any, eq(true)).returns(Future.successful(DeploymentPlan.empty))
+
+      val podSpecJsonWithPersistentVolume =
+        """
+          | { "id": "/foo/mypod",
+          |   "role": "foo",
+          |   "containers": [ {
+          |     "name": "dataapp",
+          |     "resources": { "cpus": 0.03, "mem": 64 },
+          |     "image": { "kind": "DOCKER", "id": "busybox" },
+          |     "exec": { "command": { "shell": "sleep 1" } },
+          |     "volumeMounts": [ { "name": "pst", "mountPath": "pst1", "readOnly": false } ]
+          |   } ],
+          |   "volumes": [ {
+          |     "name": "pst",
+          |     "persistent": { "type": "root", "size": 10 }
+          |   } ] }
+        """.stripMargin
+
+      val response = asyncRequest { r =>
+        f.podsResource.update("/foo/mypod", podSpecJsonWithPersistentVolume.getBytes(), force = true, f.auth.request, r)
+      }
+      withClue(s"response body: ${response.getEntity}") {
+        response.getStatus should be(HttpServletResponse.SC_OK)
       }
     }